Change PersonId return type from int to long

+++ src/Modules/Stakeholders/Explorer.Stakeholders.Infrastructure/Authentication/ClaimsPrincipalExtensions.cs
@@ -4,6 +4,6 @@ namespace Explorer.Stakeholders.Infrastructure.Authentication;
 
 public static class ClaimsPrincipalExtensions
 {
-    public static int PersonId(this ClaimsPrincipal user)
-        => int.Parse(user.Claims.First(i => i.Type == "personId").Value);
+    public static long PersonId(this ClaimsPrincipal user)
+        => long.Parse(user.Claims.First(i => i.Type == "personId").Value);
 }
